automatically backed up by learn

+++ lib/bachelor.rb
@@ -61,7 +61,7 @@ def get_average_age_for_season(data, season)
   i = 0
   avg = 0
   while i < age.length
-    avg += age[i].to_i
+    avg += age[i].to_f
     i += 1
   end
   avg = avg.to_f / age.length.to_f
